// InvoicerClient/src/app/api/model/getDashboardStatsResponse.ts
/**
 * Invoicer API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InvoiceStatusSummary } from './invoiceStatusSummary';
import { ClientIncomeItem } from './clientIncomeItem';
import { MonthlyIncomeItem } from './monthlyIncomeItem';

export interface GetDashboardStatsResponse {
  monthlyIncome?: Array<MonthlyIncomeItem> | null;
  incomeByClient?: Array<ClientIncomeItem> | null;
  statusSummary?: InvoiceStatusSummary;
}